Initial reorg of project and bringing old code up to Python standards
