+++ app/main/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+
+
+def create_app():
+    app = Flask(__name__)
+    return app
+
